Add files via upload
Code and exemplary .xml file. This is the first version of the code and I anticipate there will be a couple of iterations. For now, the code only reads in one .xml file and calculates policy costs to 2050. The first change will be calculating costs to 2100. Future changes will include reading in multiple .xml files.